Introduce .zshenv and move NVM inisitalisation there
Within vim, the shell isn't ran under interactive mode. I tried this,
to allow it to source .zshrc, but it kept suspending vim.
The correct way is to source using .zshenv instead. Therefore I have
moved NVM initialisation there which it what I wanted to, at this
moment, use npm and node and the likes.