Handle zsh history and install GPG
Started with me wanting to use
https://github.com/wulfgarpro/history-sync - but where was I going to
commit it? Publically? Even if it is encrypted, I have some qualms...

Instead, i'll keep GPG. It's handy and will use it in the future. For
zsh history backing up? Simple, change the location to a folder, and
then synchronise that folder using google drive.